Move remaining to test_headers
